arch-power: Remove pre-execute zero register update
The Power architecture does not define a zero register but
the simulator expects one and always sets it to zero in the
pre-execute stage of an instruction. This update shows up
in the logs when running the simulator with debug flags and
can be confusing as no instructions actually modify it.

This removes the redundant updates for this architecture.

Change-Id: I5e449f0144fcf4fbcaa15aed8166f132aa17bb10
Signed-off-by: Sandipan Das <sandipan@linux.ibm.com>